reorg FVGQ codes
