+++ test/runtests.jl
@@ -16,10 +16,10 @@ include("make_perturbation_model.jl")
 include("first_order_perturbation.jl")
 include("second_order_perturbation.jl")
 # include("first_order_sequence.jl")
-include("sequence_DifferenceEquations.jl")
+# include("sequence_DifferenceEquations.jl")
 # include("second_order_sequence.jl")
 # include("rbc_estimation.jl")
 include("sgu.jl")
-# include("FVGQ20.jl")
+include("FVGQ20.jl")
 include("symbolic_utils.jl")
 include("utils.jl")
